+++ sr_port/bml_find_free.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2007 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -33,35 +33,44 @@ int4 bml_find_free(int4 hint, uchar_ptr_t base_addr, int4 total_bits, bool *used
 		hint = 0;
 	for (ptr = base_addr + hint, top = base_addr + (total_bits +7) / 8 - 1; ptr < top; ptr++)
 	{	if (*ptr)
-		{	if (*ptr > 63)
-			{	if (*ptr > 127)
+		{
+			if (*ptr > 63)
+			{
+				if (*ptr > 127)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 3;
-			}else if (*ptr > 15)
-			{	if (*ptr > 31)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 3 );
+			}
+			else if (*ptr > 15)
+			{
+				if (*ptr > 31)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 2;
-			}else if (*ptr > 3)
-			{	if (*ptr > 7)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 2 );
+			}
+			else if (*ptr > 3)
+			{
+				if (*ptr > 7)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 1;
-			}else
-			{	if (*ptr == 2)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 1 );
+			}
+			else
+			{
+				if (*ptr == 2)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK);
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) );
 			}
 		}
 	}
 	if (*ptr)	/* Special processing for last byte as may be only partially valid */
-	{	bits = total_bits % 8;
+	{
+		bits = total_bits % 8;
 		if (bits == 6)
 			valid = *ptr & 63;
 		else if (bits == 4)
@@ -71,59 +80,75 @@ int4 bml_find_free(int4 hint, uchar_ptr_t base_addr, int4 total_bits, bool *used
 		else
 			valid = *ptr;
 		if (valid)
-		{	if (valid > 63)
-			{	if (valid > 127)
+		{
+			if (valid > 63)
+			{
+				if (valid > 127)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 3;
-			}else if (valid > 15)
-			{	if (valid > 31)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 3 );
+			}
+			else if (valid > 15)
+			{
+				if (valid > 31)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 2;
-			}else if (valid > 3)
-			{	if (valid > 7)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 2 );
+			}
+			else if (valid > 3)
+			{
+				if (valid > 7)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 1;
-			}else
-			{	if (valid == 2)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 1 );
+			}
+			else
+			{
+				if (valid == 2)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK);
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) );
 			}
 		}
 	}
 	for (ptr = base_addr, top = base_addr + hint; ptr < top; ptr++)
-	{	if (*ptr)
-		{	if (*ptr > 63)
-			{	if (*ptr > 127)
+	{
+		if (*ptr)
+		{
+			if (*ptr > 63)
+			{
+				if (*ptr > 127)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 3;
-			}else if (*ptr > 15)
-			{	if (*ptr > 31)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 3 );
+			}
+			else if (*ptr > 15)
+			{
+				if (*ptr > 31)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 2;
-			}else if (*ptr > 3)
-			{	if (*ptr > 7)
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 2 );
+			}
+			else if (*ptr > 3)
+			{
+				if (*ptr > 7)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK) + 1;
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) + 1 );
 			}else
-			{	if (*ptr == 2)
+			{
+				if (*ptr == 2)
 					*used = TRUE;
 				else
 					*used = FALSE;
-				return (ptr - base_addr) * (8/BML_BITS_PER_BLK);
+				return (int4)((ptr - base_addr) * (8 / BML_BITS_PER_BLK) );
 			}
 		}
 	}
