+++ sr_port/dse_data.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2007 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -70,7 +70,7 @@ int dse_data(char *dst, int *len)
 		else
 			*dst++ = *src;
 	}
-	*len = dst - bot;
+	*len = (int)(dst - bot);
 	return TRUE;
 
 }
