+++ sr_port/dse_order.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2005 Fidelity Information Services, Inc	*
+ *	Copyright 2001, 2007 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -108,7 +108,7 @@ int dse_order(block_id srch,
 				break;
 		}
 	}
-	*op = rp - bp;
+	*op = (int4)(rp - bp);
 	if (*pp == patch_find_blk && !dir_data_blk)
 	{
 		patch_left_sib = last;
