+++ sr_port/dse_rmrec.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2005 Fidelity Information Services, Inc	*
+ *	Copyright 2001, 2007 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -134,7 +134,7 @@ void dse_rmrec(void)
 			{
 				if (((blk_hdr_ptr_t) lbp)->levl)
 					util_out_print("Warning:  removed a star record from the end of this block.", TRUE);
-				((blk_hdr_ptr_t)lbp)->bsiz = rp_base - lbp;
+				((blk_hdr_ptr_t)lbp)->bsiz = (unsigned int)(rp_base - lbp);
 				BLK_INIT(bs_ptr, bs1);
 				BLK_SEG(bs_ptr, (uchar_ptr_t)lbp + sizeof(blk_hdr),
 					(int)((blk_hdr_ptr_t)lbp)->bsiz - sizeof(blk_hdr));
@@ -185,7 +185,7 @@ void dse_rmrec(void)
 		PUT_SHORT(&((rec_hdr_ptr_t)rp_base)->rsiz, rsize);
 		memcpy(rp_base + sizeof(rec_hdr), &patch_comp_key[i], patch_comp_count - i);
 		memcpy(rp_base + sizeof(rec_hdr) + patch_comp_count - i, key_top, b_top - key_top);
-		((blk_hdr_ptr_t)lbp)->bsiz = rp_base + rsize - lbp + b_top - r_top;
+		((blk_hdr_ptr_t)lbp)->bsiz = (unsigned int)(rp_base + rsize - lbp + b_top - r_top);
 		BLK_INIT(bs_ptr, bs1);
 		BLK_SEG(bs_ptr, (uchar_ptr_t)lbp + sizeof(blk_hdr), ((blk_hdr_ptr_t)lbp)->bsiz - sizeof(blk_hdr));
 		if (!BLK_FINI(bs_ptr, bs1))
